requirements.txt
numpy==1.19.5
pandas==1.3.2
imagesize==1.2.0
lxml==4.6.3
opencv-python==4.5.3.56
Pillow==8.3.2
